Update to version 1.41
- Files are now validated when loading sprite images. (Make sure they
are in png format, all other formats will be ignored) This fixed the
"thumbs.db" crash.
- Updated the UI with more animations, new style to progressBar, also
text over progress bar indicating files completed.
- Fixed several bugs I found looking through the source
- You can now name images. (The names are stored in the archives when
you write them, and loaded when you load the archives)
- By default, if a sprite has no name the index will be displayed. If
the sprite has a name, the name will be shown rather than the index.
- You can now filter the list based on the index of the sprite, or the
sprites name (Before you could only search the index of the sprite not
the name)
- ImageView is now capped at a width of either 512 or height of 512. If
the image you select from the list has a larger width or height than 512
the ImageView will be re-sized while maintaining the images aspect
ratio. (This has no effect on the image when you write them) (This fixed
the issue of large images making the program extremely big)